+++ src/Util/GlideUrl.php
@@ -10,6 +10,7 @@ use AmpedWeb\GlideInABox\Traits\Effects;
 use AmpedWeb\GlideInABox\Traits\Encode;
 use AmpedWeb\GlideInABox\Traits\Flip;
 use AmpedWeb\GlideInABox\Traits\Orientation;
+use AmpedWeb\GlideInABox\Traits\PixelDensity;
 use AmpedWeb\GlideInABox\Traits\Size;
 use Illuminate\Support\Str;
 use League\Glide\Urls\UrlBuilder;
@@ -17,7 +18,7 @@ use League\Glide\Urls\UrlBuilderFactory;
 
 class GlideUrl
 {
-    use Orientation, Flip, Crop, Size, Adjustments, Effects, Encode;
+    use Orientation, Flip, Crop, Size, PixelDensity, Adjustments, Effects, Encode;
 
     /**
      * The filepath of our image being manipulated